3d LTS working!
3d LTS is now working too!
The confidence intervals in 3d solvers are still implemented with monte carlo sampling, rather than numerical solutions to the error ellipsoid eExtrm and eVec min and max values. Also, I am not certain that the correction factors and chi2 values are all valid for the 3d cases. The confidence intervals seem a bit strange.